App/frontend/requirements.txt
streamlit==1.40.2
requests==2.32.3
python-docx==1.1.2